+++ CHANGELOG.md
@@ -5,6 +5,9 @@
  * Added dependency arrows. The class has a dependency on the return type of the method with no public arguments.
  * Changed the output class name to include the package name.
 
+### Bug fix
+ * fix to failed to parse NullableType method parameter.
+
 ## v0.0.1 (2021-11-11)
 
 ### Added